Fixed issues with typechecking for tally aggregation
